Move files to screenspace lighting and move Reflection enum in it
