+++ com.unity.render-pipelines.high-definition/HDRP/Debug/DebugDisplay.cs
@@ -33,7 +33,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
     {
         // Used to debug SSRay model
         // 1x32 bits
-        public Lit.ProjectionModel tracingModel;
+        public ScreenSpaceLighting.ProjectionModel tracingModel;
 
         // 6x32 bits
         public uint loopStartPositionSSX;                           // Proxy, HiZ, Linear
@@ -50,7 +50,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         public float iterationLinearDepthBufferMin;                 // HiZ, Linear
         public float iterationLinearDepthBufferMax;                 // HiZ, Linear
         public float iterationLinearDepthBufferMinThickness;        // HiZ, Linear
-        public Lit.HiZIntersectionKind iterationIntersectionKind;   // HiZ
+        public ScreenSpaceLighting.HiZIntersectionKind iterationIntersectionKind;   // HiZ
         public uint iterationCellSizeW;                             // HiZ, Linear
         public uint iterationCellSizeH;                             // HiZ, Linear
         public EnvShapeType proxyShapeType;                         // Proxy
@@ -125,7 +125,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         public static GUIContent[] debuggedAlgorithmStrings = null;
         public static int[] debuggedAlgorithmValues = null;
 
-        Lit.ProjectionModel m_LastProjectionModel = Lit.ProjectionModel.None;
+        ScreenSpaceLighting.ProjectionModel m_LastProjectionModel = ScreenSpaceLighting.ProjectionModel.None;
         ScreenSpaceTracingDebug m_ScreenSpaceTracingDebugData;
         public ScreenSpaceTracingDebug screenSpaceTracingDebugData
         {
@@ -136,10 +136,10 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                 if (m_LastProjectionModel != m_ScreenSpaceTracingDebugData.tracingModel)
                 {
                     m_LastProjectionModel = m_ScreenSpaceTracingDebugData.tracingModel;
-                    RefreshScreenSpaceTracingDebug<Lit.ProjectionModel>(null, m_LastProjectionModel);
+                    RefreshScreenSpaceTracingDebug<ScreenSpaceLighting.ProjectionModel>(null, m_LastProjectionModel);
                 }
 
-                if (m_ScreenSpaceTracingDebugData.tracingModel == Lit.ProjectionModel.Proxy)
+                if (m_ScreenSpaceTracingDebugData.tracingModel == ScreenSpaceLighting.ProjectionModel.Proxy)
                 {
                     showSSRayDepthPyramid = false;
                     showSSRayGrid = false;
@@ -190,10 +190,10 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             debugScreenSpaceTracingProxyValues = debugScreenSpaceTracingProxyValueList.ToArray();
             debugScreenSpaceTracingLinearStrings = debugScreenSpaceTracingLinearStringsList.ToArray();
             debugScreenSpaceTracingLinearValues = debugScreenSpaceTracingLinearValueList.ToArray();
-            debuggedAlgorithmStrings = Enum.GetNames(typeof(Lit.ProjectionModel))
+            debuggedAlgorithmStrings = Enum.GetNames(typeof(ScreenSpaceLighting.ProjectionModel))
                 .Select(t => new GUIContent(t))
                 .ToArray();
-            debuggedAlgorithmValues = (int[])Enum.GetValues(typeof(Lit.ProjectionModel));
+            debuggedAlgorithmValues = (int[])Enum.GetValues(typeof(ScreenSpaceLighting.ProjectionModel));
         }
 
         public int GetDebugMaterialIndex()
@@ -444,7 +444,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 
                 switch (screenSpaceTracingDebugData.tracingModel)
                 {
-                    case Lit.ProjectionModel.Proxy:
+                    case ScreenSpaceLighting.ProjectionModel.Proxy:
                     {
                         debugSettingsContainer.children.Add(
                             new DebugUI.EnumField { displayName = "Debug Mode", getter = GetDebugLightingSubMode, setter = SetScreenSpaceTracingDebugMode, enumNames = debugScreenSpaceTracingProxyStrings, enumValues = debugScreenSpaceTracingProxyValues, onValueChanged = RefreshScreenSpaceTracingDebug }
@@ -468,7 +468,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                             );
                         break;
                     }
-                    case Lit.ProjectionModel.HiZ:
+                    case ScreenSpaceLighting.ProjectionModel.HiZ:
                     {
                         debugSettingsContainer.children.Insert(1, new DebugUI.Value { displayName = string.Empty, getter = () => "Press PageUp/PageDown to Increase/Decrease the HiZ step." });
                         debugSettingsContainer.children.Add(
@@ -515,7 +515,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                             );
                         break;
                     }
-                    case Lit.ProjectionModel.Linear:
+                    case ScreenSpaceLighting.ProjectionModel.Linear:
                     {
                         debugSettingsContainer.children.Add(
                             new DebugUI.EnumField { displayName = "Debug Mode", getter = GetDebugLightingSubMode, setter = SetScreenSpaceTracingDebugMode, enumNames = debugScreenSpaceTracingLinearStrings, enumValues = debugScreenSpaceTracingLinearValues, onValueChanged = RefreshScreenSpaceTracingDebug },
