--- com.unity.render-pipelines.high-definition/HDRP/Lighting/Reflection/ScreenSpaceReflection.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 384c4d03a551c44448145f4093304119
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
