+++ com.unity.render-pipelines.high-definition/HDRP/Lighting/ScreenSpaceLighting/ScreenSpaceLighting.cs.hlsl
@@ -1,0 +1,24 @@
+//
+// This file was automatically generated. Please don't edit by hand.
+//
+
+#ifndef SCREENSPACELIGHTING_CS_HLSL
+#define SCREENSPACELIGHTING_CS_HLSL
+//
+// UnityEngine.Experimental.Rendering.HDPipeline.ScreenSpaceLighting+ProjectionModel:  static fields
+//
+#define PROJECTIONMODEL_NONE (0)
+#define PROJECTIONMODEL_PROXY (1)
+#define PROJECTIONMODEL_HI_Z (2)
+#define PROJECTIONMODEL_LINEAR (3)
+#define PROJECTIONMODEL_COUNT (4)
+
+//
+// UnityEngine.Experimental.Rendering.HDPipeline.ScreenSpaceLighting+HiZIntersectionKind:  static fields
+//
+#define HIZINTERSECTIONKIND_NONE (0)
+#define HIZINTERSECTIONKIND_CELL (1)
+#define HIZINTERSECTIONKIND_DEPTH (2)
+
+
+#endif
