+++ com.unity.render-pipelines.high-definition/HDRP/Lighting/ScreenSpaceLighting/ScreenSpaceRefraction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5e17fad69ea181b4483974138b566975
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
