+++ com.unity.render-pipelines.high-definition/HDRP/Lighting/ScreenSpaceLighting/ScreenSpaceTracing.hlsl.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: aee64cd8304a9fa46a29f5533aa77b29
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
